requirements.txt
Django==1.10.5
djangorestframework==3.5.3
olefile==0.44
Pillow==4.0.0
sorl-thumbnail==12.3




